Fixed and tested post

+++ src/main/java/com/phoebussoftware/technicalTest/DTO/CustomerDTO.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class CustomerDTO {
-  Long customerId;
+  Integer customerId;
   List<AccountDTO> accountDTOS;
   @NotNull
   @NotEmpty
